deployed forta underlying asset autotask
