+++ defender/downloaded/Forta_Underlying_Asset.js
@@ -1,127 +1,66 @@
-/* eslint-disable import/no-extraneous-dependencies */
+// Set the name of the Secret set in Autotask
+const discordSecretName = 'SecurityAlertsDiscordUrl';
+
 const axios = require('axios');
-/* eslint-enable import/no-extraneous-dependencies */
 
-const fortaApiEndpoint = 'https://api.forta.network/graphql';
-
-function createDiscordMessage(cTokenSymbol, transactionHash) {
-  // // construct the Etherscan transaction link
-  const etherscanLink = `[TX](<https://etherscan.io/tx/${transactionHash}>)`;
-
-  return `${etherscanLink} 🆙 Underlying asset for the **${cTokenSymbol}** cToken contract was upgraded`;
+async function post(url, method, headers, data) {
+  return axios({
+    url,
+    method,
+    headers,
+    data,
+  });
 }
 
-function getRandomInt(min, max) {
-  return Math.floor((Math.random() * (max - min)) + min);
-}
-
-// post to discord
 async function postToDiscord(url, message) {
+  const method = 'post';
   const headers = {
     'Content-Type': 'application/json',
   };
-
-  const body = {
-    content: message,
-  };
-
-  const discordObject = {
-    url,
-    method: 'post',
-    headers,
-    data: body,
-  };
+  const data = JSON.stringify({ content: message });
 
   let response;
   try {
     // perform the POST request
-    response = await axios(discordObject);
+    response = await post(url, method, headers, data);
   } catch (error) {
-    // is this a "too many requests" error (HTTP status 429)
+    // check if this is a "too many requests" error (HTTP status 429)
     if (error.response && error.response.status === 429) {
-      // rate-limited, retry
-      // after waiting a random amount of time between 2 and 15 seconds
-      const delay = getRandomInt(2000, 15000);
+      // the request was made and a response was received
+      // try again after waiting 5 - 50 seconds, if retry_after value is received, use that.
+      let timeout;
+      // Discord Webhook API defaults to v6, and v6 returns retry_after value in ms. Later versions
+      // use seconds, so this will need to be updated when Discord changes their default API version
+      // Ref: https://discord.com/developers/docs/reference
+      if (error.response.data
+        && error.response.data.retry_after
+        && error.response.data.retry_after < 50000) {
+        // Wait the specified amount of time + a random number to reduce
+        // overlap with newer requests. Initial testing reveals that the Discord Webhook allows 5
+        // requests and then resets the counter after 2 seconds. With a 15 second range of 5-20,
+        // this function can reliably can handle batches of 15 requests. Increase the max variable
+        // below if you anticipate a larger number of requests.
+        // Ref: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+        const min = 5000;
+        const max = 30000;
+        timeout = Math.floor(Math.random() * (max - min) + min);
+        timeout += error.response.data.retry_after;
+      } else {
+        // If retry_after is larger than 50 seconds, then just wait 50 seconds.
+        timeout = 50000;
+      }
       // eslint-disable-next-line no-promise-executor-return
-      const promise = new Promise((resolve) => setTimeout(resolve, delay));
+      const promise = new Promise((resolve) => setTimeout(resolve, timeout));
       await promise;
-      response = await axios(discordObject);
+      response = await post(url, method, headers, data);
     } else {
       // re-throw the error if it's not from a 429 status
       throw error;
     }
   }
-
   return response;
 }
 
-async function getFortaAlerts(botId, transactionHash) {
-  const headers = {
-    'content-type': 'application/json',
-  };
-
-  // construct query with data that you want to get back
-  const graphqlQuery = {
-    operationName: 'recentAlerts',
-    query: `query recentAlerts($input: AlertsInput) {
-      alerts(input: $input) {
-        pageInfo {
-          hasNextPage
-          endCursor {
-            alertId
-            blockNumber
-          }
-        }
-        alerts {
-          createdAt
-          name
-          protocol
-          findingType
-          hash
-          source {
-            transactionHash
-            block {
-              number
-              chainId
-            }
-            bot {
-              id
-            }
-          }
-          severity
-          metadata
-          description
-        }
-      }
-    }`,
-    variables: {
-      input: {
-        first: 100,
-        bots: [botId],
-        transactionHash,
-        createdSince: 0,
-        chainId: 1,
-      },
-    },
-  };
-
-  // perform the POST request
-  const response = await axios.post(
-    fortaApiEndpoint,
-    graphqlQuery,
-    headers,
-  );
-
-  const { data } = response;
-  if (data === undefined) {
-    return undefined;
-  }
-
-  const { data: { alerts: { alerts } } } = data;
-  return alerts;
-}
-
-// entry point for autotask
 // eslint-disable-next-line func-names
 exports.handler = async function (autotaskEvent) {
   // ensure that the autotaskEvent Object exists
@@ -129,18 +68,34 @@ exports.handler = async function (autotaskEvent) {
     throw new Error('autotaskEvent undefined');
   }
 
-  const { secrets, request } = autotaskEvent;
+  const { secrets } = autotaskEvent;
   if (secrets === undefined) {
     throw new Error('secrets undefined');
   }
 
-  // ensure that there is a DiscordUrl secret. Name changes depending on what webhook secret you use
-  const { SecurityAlertsDiscordUrl: discordUrl } = secrets;
+  // ensure that there is a DiscordUrl secret
+  const discordUrl = secrets[discordSecretName];
   if (discordUrl === undefined) {
-    throw new Error('SecurityAlertsDiscordUrl undefined');
+    throw new Error('discordUrl undefined');
+  }
+
+  // Ref: https://developer.mozilla.org/en-US/docs/Web/API/URL
+  function isValidUrl(string) {
+    let url;
+    try {
+      url = new URL(string);
+    } catch (_) {
+      return false;
+    }
+    return url.href;
+  }
+
+  if (isValidUrl(discordUrl) === false) {
+    throw new Error('discordUrl is not a valid URL');
   }
 
   // ensure that the request key exists within the autotaskEvent Object
+  const { request } = autotaskEvent;
   if (request === undefined) {
     throw new Error('request undefined');
   }
@@ -157,47 +112,40 @@ exports.handler = async function (autotaskEvent) {
     throw new Error('alert undefined');
   }
 
-  // extract the transaction hash and bot ID from the alert Object
+  // ensure that the alert key exists within the body Object
+  const { source } = body;
+  if (source === undefined) {
+    throw new Error('source undefined');
+  }
+
+  // extract the metadata from the alert Object
+  const { metadata } = alert;
+  if (metadata === undefined) {
+    throw new Error('metadata undefined');
+  }
+
+  // extract the hashes from the source Object
   const {
-    hash,
-    source: {
-      transactionHash,
-      agent: {
-        id: botId,
-      },
-    },
-  } = alert;
+    transactionHash,
+  } = source;
 
-  // retrieve the metadata from the Forta public API
-  let alerts = await getFortaAlerts(botId, transactionHash);
-  alerts = alerts.filter((alertObject) => alertObject.hash === hash);
+  // Start of usual modifications to the autotask script
+  // extract the metadata
+  const {
+    cTokenSymbol,
+  } = metadata;
+  if (cTokenSymbol === undefined) {
+    throw new Error('cTokenSymbol undefined');
+  }
 
-  const promises = alerts.map((alertData) => {
-    const { metadata } = alertData;
-    const { cTokenSymbol } = metadata;
+  const etherscanLink = `[TX](<https://etherscan.io/tx/${transactionHash}>)`;
 
-    return createDiscordMessage(
-      cTokenSymbol,
-      transactionHash,
-    );
-  });
-
-  // wait for the promises to settle
-  let results = await Promise.allSettled(promises);
+  const message = `${etherscanLink} 🆙 Underlying asset for the **${cTokenSymbol}** cToken contract was upgraded`;
 
   // create promises for posting messages to Discord webhook
-  const discordPromises = results.map((result) => {
-      console.log(`${result.value}`);
-      return postToDiscord(discordUrl, `${result.value}`);
-  });
-
-  // wait for the promises to settle
-  results = await Promise.allSettled(discordPromises);
-  results = results.filter((result) => result.status === 'rejected');
-
-  if (results.length > 0) {
-    throw new Error(results[0].reason);
-  }
+  // with Log Forwarding enabled, this console.log will forward the text string to Dune Analytics
+  console.log(message);
+  await postToDiscord(discordUrl, message);
 
   return {};
 };
