+++ defender/downloaded/Forta_Underlying_Asset.spec.js
@@ -56,7 +56,7 @@ if (discordSecretName !== discordEnvSecretName) {
 }
 
 // eslint-disable-next-line import/no-useless-path-segments
-const { handler } = require('../downloaded/Forta_Underlying_Asset');
+const { handler } = require('./Forta_Underlying_Asset');
 
 function createFortaSentinelEvent(finding, blockHash, txHash) {
   // Generally findings go from the Bot, to Scan Node, to Sentinel, to Autotasks
